Fix meal image analysis reasoning effort and harden coach image flow.
Replace unsupported gpt-5.4-nano reasoning effort with validated low effort, add model config guardrails, and surface backend error categories for gateway failures.

+++ Fitness Coach/Application/StateBuilders/Coach/CoachHealthIntelligenceSnapshotLoader.swift
@@ -90,6 +90,6 @@ enum CoachHealthIntelligenceSnapshotLoader {
 
     private static func timeoutMilliseconds(_ timeout: Duration) -> Int {
         let components = timeout.components
-        return components.seconds * 1_000 + components.attoseconds / 1_000_000_000_000_000
+        return Int(components.seconds * 1_000 + components.attoseconds / 1_000_000_000_000_000)
     }
 }
